requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
tensorflow==2.13.0
Pillow==10.0.0
numpy==1.24.3
